LayoutTests/http/tests/site-isolation/inspector/page/resource-tree-cross-origin-iframe.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<!DOCTYPE html>
<script src="../../../inspector/resources/inspector-test.js"></script>
<script>
function addCrossOriginIFrame() {
    let crossOriginHost = location.hostname === "localhost" ? "127.0.0.1" : "localhost";
    let iframe = document.createElement("iframe");
    iframe.name = "child-frame";
    iframe.src = `http://${crossOriginHost}:8000/site-isolation/inspector/page/resources/resource-tree-frame.html`;
    return new Promise(resolve => {
        iframe.addEventListener("load", () => resolve(), { once: true });
        document.body.appendChild(iframe);
    });
}

function test() {
    let suite = InspectorTest.createAsyncSuite("SiteIsolation.Page.GetResourceTreeCrossOriginIFrame");

    suite.addTestCase({
        name: "SiteIsolation.Page.GetResourceTreeCrossOriginIFrame.MainFrameAndChild",
        description: "getResourceTree should return the main frame plus a cross-origin child frame with parentId/name/origin populated.",
        async test() {
            let targetAddedPromise = WI.targetManager.awaitEvent(WI.TargetManager.Event.TargetAdded);
            // Add the cross-origin iframe from within the test so we can await the
            // frame target it spawns in a separate WebContent process. The child
            // commits in its own process, so the inspectedPage's WebFrameProxy never
            // observes its didCommitLoadForFrame; the frame target appearing is the
            // reliable signal that the cross-process frame exists.
            await InspectorTest.evaluateInPage("addCrossOriginIFrame()");
            await targetAddedPromise;

            // Under Site Isolation, use the multiplexing backendTarget because
            // its PageAgent is the UIProcess proxy that sees every WebContent
            // process. The per-page target's PageAgent (== window.PageAgent)
            // would only return the local frame's view.
            let { frameTree } = await WI.backendTarget.PageAgent.getResourceTree();

            InspectorTest.expectThat(!!frameTree.frame.id, "Main frame should have a non-empty id.");
            InspectorTest.expectEqual(frameTree.frame.parentId, undefined, "Main frame should have no parentId.");
            InspectorTest.expectThat(/127\.0\.0\.1|localhost/.test(frameTree.frame.url), "Main frame URL should be on the test host.");
            InspectorTest.expectThat(/127\.0\.0\.1|localhost/.test(frameTree.frame.securityOrigin), "Main frame security origin should be on the test host.");

            InspectorTest.expectEqual(frameTree.childFrames?.length, 1, "Should have exactly one child frame in the tree.");

            let child = frameTree.childFrames[0];
            InspectorTest.expectThat(!!child.frame.id, "Child frame should have a non-empty id.");
            InspectorTest.expectThat(child.frame.id !== frameTree.frame.id, "Child frame id should differ from main frame id.");
            InspectorTest.expectEqual(child.frame.parentId, frameTree.frame.id, "Child frame's parentId should match main frame's id.");
            InspectorTest.expectEqual(child.frame.name, "child-frame", "Child frame's name attribute should be reported.");
            // FIXME: <https://webkit.org/b/308896> The inspectedPage's WebFrameProxy never
            // observes the cross-origin child's didCommitLoadForFrame, so its url() and
            // documentSecurityOriginData() remain stale. Re-enable these once cross-process
            // WebFrameProxy state propagation lands.
            // InspectorTest.expectThat(child.frame.url.endsWith("/resource-tree-frame.html"), "Child frame URL should point at the iframe document.");
            // InspectorTest.expectThat(child.frame.securityOrigin !== frameTree.frame.securityOrigin, "Child frame should have a different (cross-origin) security origin from the main frame.");
        }
    });

    suite.runTestCasesAndFinish();
}
</script>
<body onload="runTest()">
<p>Test that Page.getResourceTree returns the cross-origin frame tree under site isolation.</p>
</body>
